move SearchStatusComponent to Developers namespace
