+++ app/components/developers/search_status_component.html.erb
